APU: Removed debugger, fixed sampling issues, fixed duty cycle in squares
Debugger is not really useful, much better to record system audio and
inspeect in external program. Sampling issues: even though the periods
are expressed in APU cycles, we apparently cannot just execute every other
frame.

This is because the period is actually X APU cycles + 1 CPU cycle since
the counter goes from period...0. So we actually need to count every
cycle in order to get it exactly correct.

Duty cycle was incorrect in the square wave generators, I used the first
sequences from the wiki, but they are expressed in the format that occurs
when the index is decremented from 0. Since I increment from 0 instead
I should use other tables.